requirements.txt
django~=3.2.0
pillow
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-restfranework-simplejwt